adds new custom filter to be able to filter by year published
added checkboxes to additionally filter by year published, since the search is really a psuedosearch, it was the next greatest specifier.
this change included: adding a working array for the years selected to be pushed to, and for the filter to check if any of those were called to use by the user.

additonally, commit adds: jquery functionality, for future styling planned.